Resolve most clang-tidy warnings

+++ firmware/ventilator-controller-stm32/Core/Inc/Pufferfish/HAL/CRC.h
@@ -26,7 +26,7 @@ namespace HAL {
  * @param xorOut    value to be xor-ed with the final output
  * @return an 8-bits CRC of the given data
  */
-uint8_t compute_crc8(const uint8_t *data, int size, uint8_t polynomial,
+uint8_t compute_crc8(const uint8_t *data, size_t size, uint8_t polynomial,
                      uint8_t init = 0x00, bool ref_in = false,
                      bool ref_out = false, uint8_t xor_out = 0x00);
